requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1
django-environ==0.9.0
django-filter==22.1
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
factory-boy==3.2.1
Faker==14.2.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
openapi-codec==1.3.2
psycopg2-binary==2.9.3
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
